refactor(dashboard-for-dapps): extract common code
