+++ apps/dashboard-for-dapps/src/hooks/index.ts
@@ -1,0 +1,6 @@
+import { useLocalStorage } from "@uidotdev/usehooks";
+
+export const useSecretKey = () => {
+  const [secretKey, setSecretKey] = useLocalStorage("SECRET_KEY", "");
+  return [secretKey, setSecretKey] as const;
+};
